perf: optimize 23 source files across katgpt-rs
- channel_simd: SIMD dot_row, fill() replacing manual loop, debug_assert
- static_cal: zero-alloc blake3::Hasher streaming for commit/verify
- newton_schulz: remove dead 'at' buffer (n*n*4 bytes saved per scratch)
- mux_demux: eliminate unwrap(), replace powi with cumulative multiply
- inference_router: extract recompile-hint helper, flatten forward_batch Vec<Vec<f32>> to Vec<f32>
- trigger_gate: expect() replacing unwrap(), inline evaluate() to avoid double mutex lock
- weights: #[repr(C)] on WeightSlice/LayerSlices, expect() replacing unwrap()
- dllm: remove let _ =, add early-exit n_mask==0 guard
- osc_kv: extract store_into_layer/dequantize_from_layer helpers (dedup 40 lines)
- targeted_precision: hoist seq_len*head_dim out of iterator
- main: zero-alloc bench index parsing (eliminate to_string_lossy().to_string())
- validator/syn_pruner: expect() replacing unwrap() on 6 mutex sites
- validator/types: #[repr(u8)] on ErrorKind enum
- octopus/types: remove #[repr(C)] from TripletIndices (6 bytes vs 8)
- percepta/types: usize→u32, i64→i32 in HullMeta
- percepta/transformer: String→&'static str for stop_token
- benchmark/mod: BenchResult field reorder, BenchCategory #[repr(u8)]
- benchmark/noise: remove let _ = on rng.f32()

+++ src/percepta/cumsum.rs
@@ -45,7 +45,7 @@ impl CumSum {
     /// Uses `kx = 1.0, ky = 0.0` so all entries share the same key,
     /// making them collinear (identical score). The `val` stores
     /// the actual value at index 0 and position at index 1.
-    pub fn insert(&mut self, value: f64, position: f64, seq: i64) {
+    pub fn insert(&mut self, value: f64, position: f64, seq: i32) {
         // Uniform key: all keys are the same, so all tie
         let key = [1.0, 0.0];
         let val = [value, position];
